Move the file use statement service to the right dir
It's mainly used by the class declaration related classes, so it belongs in that directory.
